♻️ Move landing page components to _components folder
