+++ src/app/_components/PageToolbar.tsx
@@ -14,7 +14,7 @@ import Icon from "@mdi/react";
 import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { Button } from "../button";
+import { Button } from "../../components/button";
 import LogOutButton from "./LogOutButton";
 
 export default function PageToolbar() {
